Add multi-arena episodes
Adds the feature of multi-arena episodes, by looking for the mergeNextArena
parameter in an arena's yaml. When a good (not timeout or death) episode end
is reached TrainingAgent checks TrainingArena to see if it should end the
episode or merge the next arena. If the latter it calls the new TrainingArena
function loadNextArena.