Add multi-arena episodes (#35)
* Add multi-arena episodes

Adds the feature of multi-arena episodes, by looking for the mergeNextArena
parameter in an arena's yaml. When a good (not timeout or death) episode end
is reached TrainingAgent checks TrainingArena to see if it should end the
episode or merge the next arena. If the latter it calls the new TrainingArena
function loadNextArena.

Also:
- Remove getArenaCount as it duplicated GetTotalArenas.
- GetConfiguration previously returned a bool and there was some special handling for the case it couldn't find a configuration. We've removed the special handling so removing the bool return to make
using it a bit less clunky.

---------

Co-authored-by: Ben Slater <benslater>