added code to save to csv file
its foundational right now and saves per step information on agent. it needs to be able to save the .csv file outside of the root directory (currently in unity folder so not ideal to user).

Next step would be to save to desktop or better still to save the .csv in a specified location in the yaml file as a new syntax.