further work on data logs:
The logs are now stored in a folder called observationLogs, in editor, play and train modes. The location of the editor/play is root directory, where in the build/training, it's currently: ""C:\Users\ia424\Desktop\WINDOWS\AnimalAI_Data\observationLogs\Observations.csv"" (on my PC).

Next would need to see if we can store the folder and .csv files one directory up or potentially in root directory. Will also need to see if there is any FPS lag as the data is being written simultaneously for obvious reasons...