the folder and .csv files are now stored corectly
for editor, in root directory of project files and;
for builds, in root directory of .exe file.

Ideally, the logic for this feature should have its own class (and will so in the future).

Tested on editor plus builds running on play and train modes. The folder and files are produced as expected.