IMPORTANT:
Updated code to better handle threading and logging to .csv file.

Before it was sort of unpredictably flushing to .csv file. now, there is a lock implemented which avoids multiple threads and flushes occuring.

Also, good to note is that an episode lasts 501 steps exactly. This is also reflected in the batch (buffersize) so now each episode's data is flushed in sync with the episode duration (101 time-frames).

Also removed a while loop which i think used too much cpu without needing it (now mor eefficient).