ok so:
So far there doesnt seem to be a bottleneck anymore as im handling the threads more gracefully (in general but more importantly during training) and using a thread lock mechanism to make sure the threads are controlled.

Also, the batched and framecountered approach seems to be working well as im no longer getting fps lags during training. this was the case as collectobservations is called every frame (by mlagents) and because were calling logtocsv() every method, this was causing a major bottleneck in CPU resources. Now, i need to make sure that the data is logged (flushed) right before training ends or another appropriate implementation as the .csv file was empty in my tests. This suggests the data cant be flushed in time.