found and fixed bug where:
sometimes there would be multiple rows with the same data and/or step values. Simply added a check to see if the preceding row had identical values - if not, log, if yes, skip that row. Unlikely that the same row will have the absolute identical values in every column as a note.