Upgrade mlagents to version 3.0.0, editor to ver 2022.3.50f1
This change bumps the version of mlagents from 2.3.0-exp.3 to the latest version 3.0.0. This required also updating the unity editor version to 2022.3.50f1 (the latest LTS version at the time of making this change) as the minimum supported version was bumped between these versions.

This change required updating a call to RayPerceptionSensor.Perceive, as
the interface for that function had been changed in a breaking way.