modified prefab.cs and goal.cs to accept yaml colors as overrides
The default colors are applied if no specified colors are present for the objects that support this new feature. Also improved performance by only creating a new material instance for the object if the color is to be changed.

Added a new method specifically for handling render-related color specifications. This is needed as the goals (rewards such as GoodGoal) have an emissive property, which must match the basecolor. If they dont, there will be a blend between default and specified yaml colors.

Added checks for HollowBox (special case here) to apply the color changes to it's child objects - otherwise the color changes would not apply.