changes:
So, a problem was occurring during testing (WEBGL) where some arenas would be skipped or overwritten. This was probably due to the merging of configurations in AAI3EnvironmentManager.cs, which was eviddent only when there were multiple yaml files in StreamingAssets/Yaml folder. The current code simplifies the implentation of a mergedconfigurations format, which now go through all arenas defined, regardless of IDs assigned (it bypasses the IDs, just simply stabalises the indexing (by assigning original, fresh ID's upon reading the yamls).

Further testing needs to be implemented or an alternative approach is required such as not merging the yaml files and their arena contents.