reverted to using non-merged yaml configurations for simplicity.
Yamls load in order per yaml file. There is an initial arena upon launching webgl where its used as a placeholder as the yamls are simply not loading quick enough, which the first arena is designed for (a frozen state). After the first dummy arena, it never occurs agian. it could be a bug or simply a requirement for this setup for aai webgl. However, it could be used to prepare the experiment for the participant (as a buffer arena before beginning the experiment).

Furthermore, all the global and local parameters are working correctly.