Bug Fix: Wait on the CSV writing thread to exit before closing the CSV writer
Previously in the CSV writer class there was a race condition when the Shutdown() function was called. If the final call to FlushLogQueue in the thread started by StartFlushThread occurred after the writer.Close in Shutdown, the thread would attempt to write to a closed writer and so throw an error. This change fixes the race condition by waiting on the writing thread before closing the writer.

As the error occurs on a different thread, this failures were previously swallowed. I haven't gone so far as to propagate these errors to the main thread, but the test Test_ShuttingDownDoesNotLogError captures the race condition by checking no errors have been thrown on the other thread. Note: the error surfaces much more often just after unity reloads after a code change, probably because of some JIT compilation initialisation stuff I don't really understand slowing down the extra thread.

Note that this change uses a dedicated thread instead of the thread pool. We don't expect this to have a big performance implication since the thread is only started once

Testing:
- New unit test
- CSV writing E2E tests pass