Bug Fix: Add 'Goals Collected' loglines to the CSV writer queue instead of using the StreamWriter directly
Logs are written to the csv file in two stages, first they are queued in a ConcurrentQueue (logQueue), and then periodically (when the length of the queue exceeds bufferSize) are flushed to a StreamWriter (writer). At the end of an episode logQueue is flushed by TrainingAgent using the FlushLogQueue method.

Previously, 'Goals Collected' lines were written directly to the StreamWriter at the start of each episode except the first and the end of the training session, with the apparent intention that they appear in between blocks of rows indicating episodes. The manual writing meant that if the episode was shorter than bufferSize loglines it would appear at the top of the episode instead of at the bottom as intended, so instead add these lines to logQueue.

Tested by adding Test_GoalsCollectedLines_ComeAtTheEndOfEpisodes, which failed before the change and now passess, and updating the CSV E2E test.