Remove onlyCloseWriter parameter from CSVWriter shutdown function
This was included as a stopgap to make sure not too many things were changed in the refactor of CSV writing into its own class.

Note: Coming back to this I'm not sure what my thinking behind including this was, as it seems to not really be used (Shutdown is only called in one place in TrainingAgent). However the code looks right, and the unit and E2E tests pass