Bug fix: Accept both t/pass_mark and timeLimit/passMark in YAML definitions
The previous change introduced a bug, because setting aliases on timeLimit and passMark meant that the original names were not accessible.

This change fixes that issue by introducing a backing variable for deserialising each of the namings to Arena, and a single variable whose getter/setter maps to the backing variables.