Allow setting the SpawnerButton Operation from the YAML configuration
Also:

Don’t name the SpawnedObjectHolder, as we can search for it by its existing name

Push spawning behaviour from the spawnOperation down into a new trainingArena.AddNewItemToArena function

Perform the operation at the moment the button is fully depressed instead of when it returns to its original position. The old behaviour was giving it an unnatural pause after pushing the button before the spawn happens. TODO: Flag this behaviour change in the release notes in case it makes anything existing weird

Attach a YAMLDefs.Item instead of a Spawnable in the legacy spawnerbutton behaviour so that its as close as possible to the new flow

Set the default spawnedRewardSize to 1 (previously this was inherited
from the prefab)